Split `kernels` out of `hosts` in `/records`.
