+++ packages/core/src/epics/kernel-lifecycle.js
@@ -31,7 +31,6 @@ import * as uuid from "uuid";
 import type { NewKernelAction, SetNotebookAction } from "../actionTypes";
 
 import type { KernelInfo } from "../records";
-import type { LocalKernelProps } from "../records/hosts";
 
 import * as selectors from "../selectors";
 import * as actions from "../actions";
