+++ packages/core/src/epics/websocket-kernel.js
@@ -14,7 +14,6 @@ import { of } from "rxjs/observable/of";
 import { from } from "rxjs/observable/from";
 import { merge } from "rxjs/observable/merge";
 
-import type { RemoteKernelProps } from "../records/hosts";
 import type { AppState } from "../records";
 
 import { kernels, shutdown } from "rx-jupyter";
